Move all exploitation-related JavaScript to their new home
