+++ data/js/detect/addons.js
