+++ lib/rex/exploitation/javascriptexploitation.rb
@@ -1,0 +1,37 @@
+# -*- coding: binary -*-
+
+require 'msf/core'
+require 'rex/text'
+require 'rex/exploitation/jsobfu'
+
+module Rex
+module Exploitation
+
+#
+# Provides exploitation functions in JavaScript
+#
+class JavascriptExploitation < JSObfu
+
+  def self.mstime_malloc
+  	js = ::File.read(::File.join(::File.dirname(__FILE__), "../", "../", "../", "data", "js", "exploitation", "mstime_malloc.js"))
+  	js = js.gsub(/W00TA/, Rex::Text.rand_text_hex(6))
+  	js = js.gsub(/W00TB/, Rex::Text.rand_text_hex(5))
+
+ 	js
+  end
+
+  def self.property_spray
+  	js = ::File.read(::File.join(::File.dirname(__FILE__), "../", "../", "../", "data", "js", "exploitation", "property_spray.js"))
+  	js = js.gsub(/sym_div_container/, Rex::Text.rand_text_alpha(rand(10) + 5))
+
+  	js
+  end
+
+  def self.heap_spray
+  	::File.read(::File.join(::File.dirname(__FILE__), "../", "../", "../", "data", "js", "exploitation", "heap_spray.js"))
+  end
+
+end
+end
+
+end
